fix some things. Add management commands for entire categories, including creating categories.
